third_party/WebKit/LayoutTests/http/tests/inspector/security/security-unknown-resource-expected.txt

Tests that requests to unresolved origins result in unknown security state and show up in the sidebar origin list.

Group: Main Origin
Group: Unknown / Canceled
<SPAN class=title >
http://unknown
</SPAN>
